3.1.3 | command - change folder strucutre, refactor ComplexCommand, fix issues
