--- src/main/java/edu/kis/powp/jobs2d/commands/OperateToCommand.java
@@ -1,18 +1,0 @@
-package edu.kis.powp.jobs2d.commands;
-
-import edu.kis.powp.jobs2d.Job2dDriver;
-
-public class OperateToCommand implements DriverCommand {
-    private int x;
-    private int y;
-
-    public OperateToCommand(int x, int y) {
-        this.x = x;
-        this.y = y;
-    }
-
-    @Override
-    public void execute(Job2dDriver driver) {
-        driver.operateTo(x, y);
-    }
-}
