+++ src/main/java/edu/kis/powp/jobs2d/drivers/command/SetPositionCommand.java
@@ -1,0 +1,17 @@
+package edu.kis.powp.jobs2d.drivers.command;
+
+import edu.kis.powp.jobs2d.Job2dDriver;
+
+public class SetPositionCommand implements DriverCommand {
+    private int x;
+    private int y;
+
+    public SetPositionCommand (int x, int y) {
+        this.x = x;
+        this.y = y;
+    }
+
+    public void execute(Job2dDriver driver) {
+        driver.setPosition(this.x, this.y);
+    }
+}
